% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{demography_filename}
\alias{demography_filename}
\title{Translates cycle years into the correct demography filename suffix,
e.g. '2001-2002' returns 'B'}
\usage{
demography_filename(year)
}
\arguments{
\item{year}{NHANES cycle, e.g. "2001-2002"}
}
\value{
suffix character e.g. "B"
}
\description{
Translates cycle years into the correct demography filename suffix,
e.g. '2001-2002' returns 'B'
}
