% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detection_frequency.R
\name{detection_frequency}
\alias{detection_frequency}
\title{Computes detection frequencies on a set of columns}
\usage{
detection_frequency(col)
}
\description{
Computes detection frequencies on a set of columns
}

