% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{download_nhanes_file}
\alias{download_nhanes_file}
\title{Download an NHANES data file from a given cycle}
\usage{
download_nhanes_file(file_name, year, destination = tempdir(), cache = TRUE,
  method = "auto")
}
\arguments{
\item{file_name}{file name}

\item{year}{NHANES cycle}

\item{destination}{directory to download the file into}

\item{cache}{whether to cache the file}

\item{method}{download method passed to download.file}
}
\value{
path to the downloaded file
}
\description{
Download an NHANES data file from a given cycle
}
