% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{file_suffix}
\alias{file_suffix}
\title{Returns the NHANES file suffix for the given year}
\usage{
file_suffix(year)
}
\arguments{
\item{year}{NHANES cycle year (e.g. "2001-2002")}
}
\value{
suffix character (e.g. "B" or "C")
}
\description{
Returns the NHANES file suffix for the given year
}

