% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import.R
\name{load_nhanes_data}
\alias{load_nhanes_data}
\title{Download NHANES data files.}
\usage{
load_nhanes_data(file_name, year, destination = tempdir(),
  demographics = FALSE, overwrite = FALSE)
}
\arguments{
\item{file_name}{NHANES file name e.g. ("EPH_E")}

\item{year}{NHANES cycle year (e.g. "2007-2008")}

\item{destination}{directory to download the files to}

\item{overwrite}{whether to overwrite the file if it already exists}

\item{merge_demographics}{auto merge demographics data into the dataset}
}
\description{
Download NHANES data files.
}
\examples{
load_nhanes_lab_data("UHG_G", "2011-2012")

load_nhanes_lab_data("HDL_E", "2007-2008", destination = "/tmp", overwrite = TRUE) # Download to /tmp directory and overwrite the file if it already exists
}

