% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{load_nhanes_description}
\alias{load_nhanes_description}
\title{Download an NHANES description file}
\usage{
load_nhanes_description(file_name, year, destination = tempdir(),
  cache = FALSE, method = "auto")
}
\arguments{
\item{file_name}{file name}

\item{year}{NHANES cycle}

\item{destination}{directory to download the file into}

\item{cache}{whether to cache the file}

\item{method}{download method passed to download.file}
}
\value{
data frame containing the file description
}
\description{
Download an NHANES description file
}
