% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze.R
\name{nhanes_analyze}
\alias{nhanes_analyze}
\title{Compute quantiles from NHANES weighted survey data}
\usage{
nhanes_analyze(analysis_fun, nhanes_data, column, comment_column = "",
  weights_column = "")
}
\arguments{
\item{analysis_fun}{function to use to analyze each variable}

\item{nhanes_data}{data frame containing NHANES data}

\item{column}{column name of the variable to compute quantiles for}

\item{comment_column}{comment column name of the variable}

\item{weights_column}{name of the weights column}
}
\value{
a data frame
}
\description{
Compute quantiles from NHANES weighted survey data
}

