% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{nhanes_cycle_years}
\alias{nhanes_cycle_years}
\title{List the valid NHANES cycle years}
\usage{
nhanes_cycle_years()
}
\value{
vector of NHANES cycle years
}
\description{
List the valid NHANES cycle years
}

