% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detection_frequency.R
\name{nhanes_detection_frequency}
\alias{nhanes_detection_frequency}
\title{Compute detection frequencies of NHANES data}
\usage{
nhanes_detection_frequency(nhanes_data, comment_columns)
}
\arguments{
\item{nhanes_data}{data frame containing NHANES data}

\item{columns}{comment column names of the variables to compute detection frequencies for}
}
\value{
named vector of detection frequencies
}
\description{
Compute detection frequencies of NHANES data
}
\examples{
dat <- load_nhanes_data("UHG_G", "2011-2012", demographics = TRUE)

# Compute detection frequency
nhanes_detection_frequency(dat, c("URXUHG"))
}

