% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_mean.R
\name{nhanes_geometric_mean}
\alias{nhanes_geometric_mean}
\title{Compute geometric means from NHANES weighted survey data}
\usage{
nhanes_geometric_mean(nhanes_data, column, weights_column = "", filter = NULL)
}
\arguments{
\item{nhanes_data}{data frame containing NHANES data}

\item{column}{column name of the variable to compute geometric means for}

\item{weights_column}{name of the weights column}

\item{filter}{logical expression used to subset the data}
}
\value{
a data frame
}
\description{
Compute geometric means from NHANES weighted survey data
}
\examples{

\dontrun{
dat <- nhanes_load_data("UHG_G", "2011-2012", demographics = TRUE)

nhanes_geometric_mean(dat, "URXUHG", "URDUHGLC", "WTSA2YR")
}

}
