% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import.R
\name{nhanes_load_demography_data}
\alias{nhanes_load_demography_data}
\title{Download NHANES demography files for a specific cycle.}
\usage{
nhanes_load_demography_data(year, destination = tempdir(),
  overwrite = FALSE)
}
\arguments{
\item{year}{NHANES cycle year (e.g. "2011-2012")}

\item{destination}{directory to download the file to}

\item{overwrite}{sets whether to overwrite the file if it already exists}
}
\description{
Download NHANES demography files for a specific cycle.
}
\examples{
nhanes_load_demography_data("2011-2012")
}

