% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantiles.R
\name{nhanes_quantiles}
\alias{nhanes_quantiles}
\title{Compute quantiles from NHANES weighted survey data}
\usage{
nhanes_quantiles(nhanes_data, column, weights_column, quantiles)
}
\arguments{
\item{nhanes_data}{data frame containing NHANES data}

\item{column}{column name of the variable to compute quantiles for}

\item{weights_column}{name of the weights column}

\item{quantiles}{numeric or vector numeric of quantiles to compute}
}
\value{
vector of quantiles
}
\description{
Compute quantiles from NHANES weighted survey data
}

