% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/variables.R
\name{nhanes_variables}
\alias{nhanes_variables}
\title{Load the NHANES comprehensive variable list}
\usage{
nhanes_variables(destination = tempfile(), cache = TRUE)
}
\arguments{
\item{destination}{where to save the variable list}

\item{cache}{whether to cache the downloaded variable list so it doesn't have to be redownloaded every time}
}
\description{
Load the NHANES comprehensive variable list
}

