% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{parse_data_files_page}
\alias{parse_data_files_page}
\title{Helper function for nhanes_data_files function}
\usage{
parse_data_files_page(component, destination = tempfile())
}
\arguments{
\item{component}{one of demographics", "dietary", "examination", "laboratory", "questionnaire"}

\item{destination}{download destination}
}
\value{
dat
}
\description{
Helper function for nhanes_data_files function
}

