% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{parse_variable_list}
\alias{parse_variable_list}
\title{Helper function for nhanes_variable function}
\usage{
parse_variable_list(component, destination = tempfile())
}
\arguments{
\item{component}{one of "Demographics", "Dietary", "Examination", "Laboratory", "Questionnaire"}

\item{destination}{download destination}
}
\value{
dat
}
\description{
Helper function for nhanes_variable function
}

