% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNHANES.R
\docType{package}
\name{RNHANES-package}
\alias{RNHANES-package}
\alias{RNHANES}
\alias{rnhanes}
\alias{rnhanes-package}
\title{RNHANES simplifies downloading and analyzing NHANES data.}
\description{
RNHANES simplifies downloading and analyzing NHANES data.
}
