% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{validate_year}
\alias{validate_year}
\title{Check that the year is in the correct format
e.g. '2001-2002' is correct and returns TRUE,
'2001' is not correct and returns FALSE}
\usage{
validate_year(year, throw_error = TRUE)
}
\arguments{
\item{year}{the year or years to validate}

\item{throw_error}{whether to throw an error if the year is invalid}
}
\description{
Check that the year is in the correct format
e.g. '2001-2002' is correct and returns TRUE,
'2001' is not correct and returns FALSE
}
