% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import.R
\name{download_nhanes_file}
\alias{download_nhanes_file}
\title{Download an NHANES data file from a given cycle}
\usage{
download_nhanes_file(file_name, year, destination = tempdir(),
  overwrite = FALSE)
}
\description{
Download an NHANES data file from a given cycle
}

