% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/variables.R
\name{parse_data_files_page}
\alias{parse_data_files_page}
\title{Helper function for nhanes_data_files function}
\usage{
parse_data_files_page(type, destination = tempfile())
}
\description{
Helper function for nhanes_data_files function
}

