% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import.R
\name{process_file_name}
\alias{process_file_name}
\title{Processes a file name to make sure it is valid and has the correct suffix and extension
File names with an extension (e.g. ".XPT") are not altered}
\usage{
process_file_name(file_name, year, extension = ".XPT")
}
\description{
Processes a file name to make sure it is valid and has the correct suffix and extension
File names with an extension (e.g. ".XPT") are not altered
}

